Extract match log
